+++ src/main/java/ch/zhaw/schiffeversenken/guicomponents/ShipHit.java
@@ -5,7 +5,7 @@ import java.awt.Graphics;
 
 public class ShipHit extends Shape {
 
-	public ShipHit(int centerX, int centerY, double width, double height, Color color) {
+	public ShipHit(double centerX, double centerY, double width, double height, Color color) {
 		super(centerX, centerY, width, height, color);
 	}
